# docs/swagger.yaml
definitions:
  models.CreateEventRequest:
    properties:
      content:
        type: string
      type:
        type: string
      user_id:
        type: string
    required:
    - content
    - type
    - user_id
    type: object
  models.UpdateEventRequest:
    properties:
      content:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
info:
  contact:
    email: rm6551@student.uni-lj.si
    name: Rok Mokotar
    url: https://www.linkedin.com/in/mokot/
  description: Event tracking documentation for the social networking app focused
    on sharing vibes.
  title: Event Tracking API Documentation
  version: "1.0"
paths:
  /events:
    get:
      description: get events
      operationId: get-events
      produces:
      - application/json
      responses: {}
      summary: List events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: create event
      operationId: create-event
      parameters:
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventRequest'
      produces:
      - application/json
      responses: {}
      summary: Create event
      tags:
      - events
  /events/{id}:
    delete:
      description: delete event
      operationId: delete-event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Delete event
      tags:
      - events
    get:
      description: get event
      operationId: get-event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Get event
      tags:
      - events
    patch:
      consumes:
      - application/json
      description: update event
      operationId: update-event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEventRequest'
      produces:
      - application/json
      responses: {}
      summary: Update event
      tags:
      - events
...
swagger: "2.0"